Remove unused entry handlers from App and cleanup imports
